--- src/pages/user/register/register-context/provider.ts
@@ -1,2 +1,0 @@
-
-}
