+++ src/pages/user/register/subPages/RegisterPw.tsx
@@ -1,6 +1,7 @@
-import React, { useEffect, useRef, useState } from "react";
-import { useUserInfo } from "./register-context/context";
-import { Modal } from "../../../components";
+import React, { useRef, useState } from "react";
+import { Modal } from "../../../../components";
+
+import { useUserInfo } from "../register-context/context";
 
 enum InfoType {
   USER_NAME,
@@ -18,54 +19,42 @@ interface Prop {
   dispatch: React.Dispatch<Action>;
 }
 
-const RegisterName = ({ dispatch }: Prop) => {
-  const { setName, userInfo } = useUserInfo();
-  const nameRef = useRef<HTMLInputElement | null>(null);
+const RegisterPw = ({ dispatch }: Prop) => {
+  const { setUserPw, userInfo } = useUserInfo();
+  const pwRef = useRef<HTMLInputElement | null>(null);
   const [modalOpened, setModalOpened] = useState(false);
   const [message, setMessage] = useState<string>("");
 
   const onNext = () => {
-    const inputName = nameRef.current?.value;
-    if (!inputName) {
-      setMessage("이름을 입력해주세요");
+    const inputPw = pwRef.current?.value;
+    if (!inputPw) {
+      setMessage("비밀번호를 입력해주세요");
       setModalOpened(true);
       return;
     }
 
-    setName(inputName);
-    dispatch({ type: InfoType.NAME });
+    setUserPw(inputPw);
+    dispatch({ type: InfoType.USER_PW });
   };
-
-  useEffect(() => {
-    nameRef.current?.focus();
-    nameRef.current?.click();
-  }, []);
 
   return (
     <section className="flex flex-col justify-between h-full">
       <div className="flex flex-col gap-10">
         <h1 className="text-start text-xl">
-          이름을
+          비밀번호를
           <br />
           입력해주세요
         </h1>
-        <div className="grid grid-cols-10 border-b-2 border-hanaGreen w-full ">
+        <div className="grid grid-cols-10 border-b-2 border-hanaGreen w-full focus:outline-none">
           <input
             className=" col-span-9 px-2 py-1 bg-transparent focus:outline-none"
-            placeholder="이름"
-            type="string"
-            ref={nameRef}
+            placeholder="비밀번호"
+            type="password"
+            ref={pwRef}
           />
           <div className="col-span-1">x</div>
         </div>
-        <div className="border-b-2 border-b-hanaSilver px-2 py-1 text-start text-hanaSilver font-extralight">
-          {Array(userInfo.userPw.length)
-            .fill(null)
-            .map((item) => {
-              return "*";
-            })}
-        </div>
-        <div className="border-b-2 border-b-hanaSilver px-2 py-1 text-start text-hanaSilver font-extralight">
+        <div className="border-b-2 border-b-hanaSilver text-hanaSilver px-2 py-1 text-start font-ha">
           {userInfo.username}
         </div>
       </div>
@@ -76,6 +65,7 @@ const RegisterName = ({ dispatch }: Prop) => {
       >
         다음
       </button>
+
       {modalOpened && (
         <div className="absolute top-0 left-0 w-full h-full flex items-center justify-center">
           <Modal
@@ -89,4 +79,4 @@ const RegisterName = ({ dispatch }: Prop) => {
   );
 };
 
-export default RegisterName;
+export default RegisterPw;
